python/chatbot/elevenlabs-chatbot-protect/pyproject.toml
[project]
name = "elevenlabs-chatbot-protect"
version = "0.1.0"
description = "ElevenLabs Voice Chatbot with Galileo Protect integration for guardrails"
requires-python = ">=3.9"
dependencies = [
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "elevenlabs>=1.0.0",
    "galileo>=1.34.0",
    "pyaudio>=0.2.14",
]

[project.scripts]
voice-chatbot = "main:main"
voice-conversation = "conversation:run_voice_conversation"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
